Module discord_util
