+++ trema.py
@@ -10,6 +10,12 @@ import discord
 
 from pymongo import MongoClient
 
+from discord_util import\
+	get_channel_by_name,\
+	member_roles_are_default,\
+	send_delayed_dm
+
+
 client = MongoClient("mongodb://root:root@localhost:27017/?authSource=admin")
 #database Trema
 mydb = client["trema"]
@@ -31,7 +37,6 @@ mycol.drop()
 ###############
 
 _DELAY_SECS_15_MIN = 15 * 60
-_ROLE_EVERYONE = "@everyone"
 
 
 def _make_arg_parser():
@@ -53,51 +58,7 @@ intents = discord.Intents.default()
 intents.members = True
 
 trema = discord.Bot(intents=intents)
-
-
-def get_channel_by_name(channel_name):
-	return discord.utils.get(trema.get_all_channels(), name=channel_name)
-
-
-def member_roles_are_default(member):
-	"""
-	Determines whether a server member does not have roles besides the default
-	ones: @everyone and the member's name.
-
-	Args:
-		member (discord.member.Member): a server member
-
-	Returns:
-		bool: True if all of member's roles are default one, False otherwise
-	"""
-	member_name = member.name
-
-	for role in member.roles:
-		role_name = role.name
-
-		if role_name != _ROLE_EVERYONE and role_name != member_name:
-			return False
-
-	return True
-
-
-async def send_delayed_dm(user, message, delay, condition=None):
-	"""
-	Sends a direct message (DM) to the specified Discord user after a delay. If
-	a condition is given, it is evaluated when the delay is up and it must
-	return True for the message to be sent.
-
-	Args:
-		user (discord.abc.User): a Discord user
-		message (str): the direct message to send to user
-		delay (int): the time to wait, in seconds, before sending the message
-		condition (function): a Boolean function that takes no argument. The
-			message is sent if it is None or it returns True. Defaults to None.
-	"""
-	await asyncio.sleep(delay)
-
-	if condition is None or condition():
-		await user.send(message)
+print(type(trema))
 
 
 @trema.event
@@ -105,7 +66,7 @@ async def on_member_join(member):
 	guild = member.guild
 	sys_chan = guild.system_channel
 	# A channel that contains integration instructions
-	instruct_chan = get_channel_by_name("accueil")
+	instruct_chan = get_channel_by_name(trema, "accueil")
 	welcome_msg =\
 		f"Heille {member.mention}!"\
 		+ f"\nBienvenue au Club CEDILLE. "\
